Updated everything

+++ HumongousResourceEditor/include/cmd/OPCodes.h
@@ -20,7 +20,7 @@ namespace resource_editor
 			command_wait
 		};
 		jump_command isJumpCode(unsigned char byte, size_t args);
-		size_t jump(ScriptInstruction& args, unsigned char* data, size_t data_size);
+		int32_t jump(ScriptInstruction& args, unsigned char* data, size_t data_size);
 
 		void extended_b_op(unsigned char* data, ArgsAllocator& return_data);
 		void extended_w_op(unsigned char* data, ArgsAllocator& return_data);
